Rename model/bpmn/xsd-json to model/bpmn/json
