+++ src/component/parser/json/BpmnJsonParser.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import BpmnModel from '../../../model/bpmn/internal/BpmnModel';
-import { BpmnJsonModel, TDefinitions } from '../../../model/bpmn/xsd-json/BPMN20';
+import { BpmnJsonModel, TDefinitions } from '../../../model/bpmn/json/BPMN20';
 import CollaborationConverter from './converter/CollaborationConverter';
 import ProcessConverter from './converter/ProcessConverter';
 import DiagramConverter from './converter/DiagramConverter';
