+++ src/component/parser/json/converter/EventDefinitionConverter.ts
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { TDefinitions } from '../../../../model/bpmn/xsd-json/BPMN20';
+import { TDefinitions } from '../../../../model/bpmn/json/BPMN20';
 import { ShapeBpmnEventKind } from '../../../../model/bpmn/internal/shape';
-import { TEventDefinition } from '../../../../model/bpmn/xsd-json/baseElement/rootElement/eventDefinition';
+import { TEventDefinition } from '../../../../model/bpmn/json/baseElement/rootElement/eventDefinition';
 import { ensureIsArray } from './ConverterUtil';
 
 export const bpmnEventKinds = Object.values(ShapeBpmnEventKind).filter(kind => {
