+++ src/component/parser/xml/BpmnXmlParser.ts
@@ -15,7 +15,7 @@
  */
 import { parse } from 'fast-xml-parser/src/parser';
 import { decodeXML } from 'entities/lib/decode';
-import { BpmnJsonModel } from '../../../model/bpmn/xsd-json/BPMN20';
+import { BpmnJsonModel } from '../../../model/bpmn/json/BPMN20';
 
 /**
  * Parse bpmn xml source
