+++ src/model/bpmn/json/baseElement/artifact.ts
