+++ src/model/bpmn/json/baseElement/conversation.ts
