+++ src/model/bpmn/json/baseElement/correlation.ts
