+++ src/model/bpmn/json/baseElement/flowElement.ts
