+++ src/model/bpmn/json/baseElement/flowNode/gateway.ts
