+++ src/model/bpmn/json/baseElement/loopCharacteristics.ts
