+++ src/model/bpmn/json/baseElement/participant.ts
