+++ src/model/bpmn/json/baseElement/resource.ts
