+++ src/model/bpmn/json/baseElement/rootElement/collaboration.ts
