+++ test/unit/component/parser/json/BpmnJsonParser.diagram.test.ts
@@ -15,7 +15,7 @@
  */
 import { parseJsonAndExpectOnlyFlowNodes, verifyShape } from './JsonTestUtils';
 import { ShapeBpmnElementKind } from '../../../../../src/model/bpmn/internal/shape/ShapeBpmnElementKind';
-import { BpmnJsonModel } from '../../../../../src/model/bpmn/xsd-json/BPMN20';
+import { BpmnJsonModel } from '../../../../../src/model/bpmn/json/BPMN20';
 
 describe('parse bpmn as json for diagram', () => {
   const jsonModelWithoutBPMNDiagram: BpmnJsonModel = {
