+++ test/unit/component/parser/json/BpmnJsonParser.messageFlow.test.ts
@@ -16,7 +16,7 @@
 import { parseJsonAndExpectOnlyEdges, verifyEdge } from './JsonTestUtils';
 import Waypoint from '../../../../../src/model/bpmn/internal/edge/Waypoint';
 import { MessageVisibleKind } from '../../../../../src/model/bpmn/internal/edge/MessageVisibleKind';
-import * as bpmndi from '../../../../../src/model/bpmn/xsd-json/BPMNDI';
+import * as bpmndi from '../../../../../src/model/bpmn/json/BPMNDI';
 
 describe('parse bpmn as json for message flow', () => {
   it(`should convert as Edge, when an message flow is an attribute (as object) of 'collaboration' (as object)`, () => {
