+++ test/unit/component/parser/json/JsonTestUtils.ts
@@ -25,7 +25,7 @@ import Label from '../../../../../src/model/bpmn/internal/Label';
 import { SequenceFlow } from '../../../../../src/model/bpmn/internal/edge/Flow';
 import { FlowKind } from '../../../../../src/model/bpmn/internal/edge/FlowKind';
 import { MessageVisibleKind } from '../../../../../src/model/bpmn/internal/edge/MessageVisibleKind';
-import { BpmnJsonModel } from '../../../../../src/model/bpmn/xsd-json/BPMN20';
+import { BpmnJsonModel } from '../../../../../src/model/bpmn/json/BPMN20';
 
 export interface ExpectedShape {
   shapeId: string;
